Split the code and fixed the first filter. #2
