+++ src/rcl/PanTableComponents/PanTable.jsx
@@ -1,10 +1,8 @@
 import {useState, useMemo, useEffect} from "react";
-import PropTypes from 'prop-types';
-import { Box, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, TableSortLabel, IconButton, Popover, TextField, Stack, Chip, Checkbox, Toolbar, Typography, Tooltip } from "@mui/material";
-import FilterListIcon from '@mui/icons-material/FilterList';
+import { Box, Paper, Table, TableBody, TableCell, TableContainer, TableRow, Stack, Chip, Checkbox, Toolbar, Typography } from "@mui/material";
 import HighlightOffIcon from '@mui/icons-material/HighlightOff';
-import { visuallyHidden } from '@mui/utils';
-import { alpha } from '@mui/material/styles';
+import EnhancedTableHead from "./EnhancedTableHead";
+import EnhancedTableToolbar from "./EnhancedTableToolbar";
 
 function descendingComparator(a, b, orderBy) {
     if (b[orderBy] < a[orderBy]) {
@@ -21,214 +19,6 @@ function getComparator(order, orderBy) {
     ? (a, b) => descendingComparator(a, b, orderBy)
     : (a, b) => -descendingComparator(a, b, orderBy);
 }
-
-function EnhancedTableHead(props) {
-    const {onSelectAllClick, numSelected, order, orderBy, rowCount, onRequestSort, columns, columnFilters, setColumnFilter, showColumnFilters, groupOperations } = props;
-    const createSortHandler = (property) => (event) => {
-        onRequestSort(event, property);
-    };
-
-    const [anchorEl, setAnchorEl] = useState(null);
-    const [activeField, setActiveField] = useState(null);
-
-    const handleFilterClick = (event, field) => {
-        event.stopPropagation();
-        setAnchorEl(event.currentTarget);
-        setActiveField(field);
-    };
-
-    const handleFilterClose = () => {
-        setAnchorEl(null);
-        setActiveField(null);
-    };
-
-    const handleFilterChange = (event) => {
-        const text = event.target.value;
-        setColumnFilter(activeField, text);
-    };
-
-    const open = Boolean(anchorEl);
-    const id = open ? 'filter-popover' : undefined;
-
-    return (
-        <TableHead>
-            <TableRow>
-                {groupOperations && 
-                <TableCell padding="checkbox">
-                    <Checkbox
-                        color="primary"
-                        indeterminate={numSelected > 0 && numSelected < rowCount}
-                        checked={rowCount > 0 && numSelected === rowCount}
-                        onChange={onSelectAllClick}
-                        slotProps={{ 'aria-label': 'select all rows' }}
-                    />
-                </TableCell>}
-                {columns.map((c) => (
-                    <TableCell
-                        key={c.field}
-                        align={c.alignRight ? 'right' : 'left'}
-                        padding={c.disablePadding ? 'none' : 'normal'}
-                        sortDirection={orderBy === c.field ? order : false}
-                    >
-                        <TableSortLabel
-                            active={orderBy === c.field}
-                            direction={orderBy === c.field ? order : 'asc'}
-                            onClick={createSortHandler(c.field)}
-                        >
-                            {c.headerName}
-                            {orderBy === c.field ? (
-                            <Box component="span" sx={visuallyHidden}>
-                                {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
-                            </Box>
-                            ) : null}
-                        </TableSortLabel>
-                        {/* Show or hide column filters */}
-                        {(showColumnFilters && !c.numeric) && 
-                            <IconButton 
-                                size="small" 
-                                onClick={(e) => handleFilterClick(e, c.field)}
-                                sx={{ ml: 1, p: 0 }}
-                                color={columnFilters[c.field] ? 'primary' : 'default'} 
-                            >
-                                <FilterListIcon fontSize="small" />
-                            </IconButton>
-                        }
-                    </TableCell>
-                ))}
-            </TableRow>
-            <Popover
-                id={id}
-                open={open}
-                anchorEl={anchorEl}
-                onClose={handleFilterClose}
-                anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
-            >
-                <Box sx={{ p: 2 }}>
-                    <TextField
-                        label={`Filter ${activeField}`}
-                        variant="outlined"
-                        value={activeField ? columnFilters[activeField] : ''}
-                        onChange={handleFilterChange}
-                        size="small"
-                        autoFocus
-                    />
-                </Box>
-            </Popover>
-        </TableHead>
-    );
-}
-  
-EnhancedTableHead.propTypes = {
-    onRequestSort: PropTypes.func.isRequired,
-    order: PropTypes.oneOf(['asc', 'desc']).isRequired,
-    orderBy: PropTypes.string.isRequired,
-    columns: PropTypes.array.isRequired,
-    setColumnFilter: PropTypes.func.isRequired,
-    columnFilters: PropTypes.object.isRequired,
-};
-
-function EnhancedTableToolbar(props) {
-    const { numSelected, selectedIds, clearSelection, tableTitle, groupOperations, dataRows } = props;
-
-    // Here we define the action handlers
-    
-    const hydratedHandlers = useMemo(() => {
-        if (groupOperations && groupOperations.length > 0) {
-            const context = {
-                selectedIds,
-                numSelected,
-                clearSelection
-            };
-
-            const handlers = {};
-
-            groupOperations.forEach(operation => {
-                // Transforms the raw array action into a complete handler so the resulting handler no longer needs arguments when executed.
-                handlers[operation.label] = () => operation.action(context, dataRows);
-            });
-
-            return handlers;
-        }
-    }, [selectedIds, numSelected, dataRows]);
-
-    return (
-        <Toolbar
-            sx={{
-                pl: { sm: 2 },
-                pr: { xs: 1, sm: 1 },
-                // Conditional styling
-                ...(numSelected > 0 && {
-                    bgcolor: (theme) =>
-                        alpha(theme.palette.primary.main, theme.palette.action.activatedOpacity),
-                }),
-            }}
-        >
-            {numSelected > 0 ? (
-                <Typography
-                    sx={{ flex: '1 1 100%' }}
-                    color="inherit"
-                    variant="subtitle1"
-                    component="div"
-                >
-                    {numSelected} selected
-                </Typography>
-            ) : (tableTitle ? (
-                <Typography
-                    sx={{ flex: '1 1 100%' }}
-                    variant="h6"
-                    id="tableTitle"
-                    component="div"
-                >
-                    {tableTitle}
-                </Typography>) : 
-                <Typography
-                    sx={{ flex: '1 1 100%' }}
-                    variant="h6"
-                    id="tableTitle"
-                    component="div"
-                >
-                    PanTable
-                </Typography>
-            )}
-
-            {numSelected > 0 ? (
-                // Action buttons
-                <>
-                    {groupOperations && groupOperations.map((operation, index) => {
-                        const IconComponent = operation.icon; 
-                        const specificHandler = hydratedHandlers[operation.label];
-
-                        return (
-                            <Tooltip title={operation.label} key={index}>
-                                <IconButton 
-                                    onClick={specificHandler} 
-                                    disabled={numSelected === 0}
-                                >
-                                    <IconComponent />
-                                </IconButton>
-                            </Tooltip>
-                        );
-                    })}
-                </>
-            ) : (
-                // Here we can add an icon for filtering if nothing is selected
-                <>
-                    {/* <Tooltip title="Filter list">
-                        <IconButton>
-                            <FilterListIcon />
-                        </IconButton>
-                    </Tooltip> */}
-                </>
-            )}
-        </Toolbar>
-    );
-}
-
-EnhancedTableToolbar.propTypes = {
-    numSelected: PropTypes.number.isRequired,
-    selectedIds: PropTypes.array.isRequired,
-    clearSelection: PropTypes.func.isRequired,
-};
 
 export default function PanTable({columns, rows, defaultFilter, setDefaultFilter, filterPreset, showColumnFilters, tableTitle, groupOperations}) {
